removing import of pathlib and moving hashing tests to gdswriter.py instead of cell.py
